big setup for api
