--- core/src/de/fhbielefeld/pmdungeon/PMDungeon.java
@@ -1,25 +1,0 @@
-package de.fhbielefeld.pmdungeon;
-
-import com.badlogic.gdx.Game;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import de.fhbielefeld.pmdungeon.screens.GameScreen;
-
-public class PMDungeon extends Game {
-
-    private SpriteBatch batch;
-
-    @Override
-    public void create() {
-        batch = new SpriteBatch();
-        this.setScreen(new GameScreen(this));
-    }
-
-    @Override
-    public void dispose() {
-        batch.dispose();
-    }
-
-    public SpriteBatch getBatch() {
-        return batch;
-    }
-}
