+++ core/src/game/PMDungeon.java
@@ -1,0 +1,24 @@
+package game;
+import com.badlogic.gdx.Game;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+
+
+public class PMDungeon extends Game {
+
+    private SpriteBatch batch;
+
+    @Override
+    public void create() {
+        batch = new SpriteBatch();
+        this.setScreen(new GameScreen(this));
+    }
+
+    @Override
+    public void dispose() {
+        batch.dispose();
+    }
+
+    public SpriteBatch getBatch() {
+        return batch;
+    }
+}
